requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
google-generativeai==0.3.2
aiohttp==3.9.1
redis==5.0.1
python-multipart==0.0.6
solders
websocket-client==1.6.4
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
